expr: Deduplicate integer constant sets.
Address sets may contain the same value multiple times in cases where
multiple string representations of the same value exist such that
duplicates cannot be prevented at the ovsdb level.
Examples of this are "10.0.0.1" and "10.0.0.1/32" or "::1" and "::01",
the former being interesting in that they do differ in that one of them
is masked and the other one is not. In both cases however they map to
the same desired flow.

When adding such a duplicate to an existing address set we currently
see this as a new address being added. If we additionally attempt to
process that change incrementally we will try to install the same
desired flow that we already installed for the original value a second
time. This previously caused the ovn-controller to crash, which is no
longer the case, but I would still consider the behavior described here
to be a bug.

Fix this by removing any duplicates from the address set before
processing it further in the ovn-controller. The simplest way to do this
appears to be to just remove any values that are "contained" in another
value in the set i.e. by a masked value that includes this value.
Conveniently, we already sort the set in a way where values that may
contain another one come right before it.
As an example ["10.0.0.1", "10.0.0.1/32", "10.0.0.0/24"] will now be
collapsed into just ["10.0.0.0/24"].

Signed-off-by: Felix Moebius <felix.moebius@mail.schwarz>